refactor(storage): lift nibbles_to_eth_compact into eth_encoding module
Move the hex-prefix (compact) path encoder out of the `ethhash`-gated
hasher and into a new unconditionally compiled `storage::eth_encoding`
module. The function is pure and has no feature-specific dependencies;
its previous gating was incidental to where it lived.

This is groundwork for an `eth_getProof`-compatible proof emitter, which
needs to emit canonical MPT-RLP node bytes regardless of which hasher
the underlying trie was built with. Forcing the emitter to live behind
`#[cfg(feature = "ethhash")]` just to reach this helper would add walls
we'd then have to tear down.

It also lines up with the broader direction of replacing the `ethhash`
compile-time feature flag with a runtime mode knob: pulling pure
encoding helpers out of the gated module shrinks what the feature
actually has to guard, so the eventual flag removal is a smaller change.

No behavior change. The unit tests move with the function and continue
to pass under both default and `--features ethhash,logger`.
